src/image_processing/requirements.txt
boto==2.40.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
Werkzeug==0.11.10
wheel==0.24.0
gcloud==0.17.0